fix all bug biginteger & cart until delete

+++ restapi/models/PostalCodeModel.py
@@ -1,4 +1,4 @@
-from sqlalchemy import Table, Column, Integer, BigInteger, String, ForeignKey
+from sqlalchemy import Table, Column, Integer, String, ForeignKey
 from config import metadata
 
 postal_code = Table('postal_codes', metadata,
@@ -6,6 +6,6 @@ postal_code = Table('postal_codes', metadata,
     Column('urban', String(100), nullable=False),
     Column('sub_district', String(100), nullable=False),
     Column('city', String(100), nullable=False),
-    Column('postal_code', BigInteger, nullable=False),
+    Column('postal_code', Integer, nullable=False),
     Column('province_code', Integer, ForeignKey('provinces.code',onupdate='cascade',ondelete='cascade'), nullable=False)
 )
